+++ src/main/java/com/gitlab/cdagaming/craftpresence/handler/TranslationHandler.java
@@ -103,12 +103,21 @@ public class TranslationHandler {
     private void getTranslationMap() {
         translationMap = Maps.newHashMap();
 
-        InputStream in = TranslationHandler.class.getResourceAsStream("/assets/"
+        InputStream in = Constants.CLASS_LOADER.getResourceAsStream("/assets/"
                 + (!StringHandler.isNullOrEmpty(modID) ? modID + "/" : "") +
                 "lang/" + languageID + (usingJSON ? ".json" : ".lang"));
-        InputStream fallbackIn = TranslationHandler.class.getResourceAsStream("/assets/"
+        InputStream fallbackIn = Constants.CLASS_LOADER.getResourceAsStream("/assets/"
                 + (!StringHandler.isNullOrEmpty(modID) ? modID + "/" : "") +
                 "lang/" + languageID.toLowerCase() + (usingJSON ? ".json" : ".lang"));
+
+        if (in == null && fallbackIn == null) {
+            in = TranslationHandler.class.getResourceAsStream("/assets/"
+                    + (!StringHandler.isNullOrEmpty(modID) ? modID + "/" : "") +
+                    "lang/" + languageID + (usingJSON ? ".json" : ".lang"));
+            fallbackIn = TranslationHandler.class.getResourceAsStream("/assets/"
+                    + (!StringHandler.isNullOrEmpty(modID) ? modID + "/" : "") +
+                    "lang/" + languageID.toLowerCase() + (usingJSON ? ".json" : ".lang"));
+        }
 
         if (in != null || fallbackIn != null) {
             BufferedReader reader = new BufferedReader(new InputStreamReader(in != null ? in : fallbackIn));